+++ config/config.js
@@ -1,5 +1,3 @@
-const path = require('path');
-const rootPath = path.normalize(__dirname + '/..');
 const env = process.env.NODE_ENV || 'development';
 
 const config = {
@@ -15,6 +13,21 @@ const config = {
     validators: [
       {
         location: 'http://localhost:3000'
+      }
+    ]
+  },
+  staging: {
+    token: {
+      totalSupply: '1000000000000000000',
+      name: 'Umbrella',
+      symbol: 'UMB'
+    },
+    chain: {
+      interval: 6
+    },
+    validators: [
+      {
+        location: 'http://ae883a71637d7493fad1b62f09469eff-1487985040.us-east-2.elb.amazonaws.com'
       }
     ]
   }