+++ scripts/deploy.js
@@ -1,62 +1,90 @@
-const bre = require("@nomiclabs/buidler");
+require('custom-env').env();
+const { INFURA_ID } = process.env;
+
+const bre = require('@nomiclabs/buidler');
 const config = require('../config/config');
 
-const currentProvider = new web3.providers.HttpProvider('http://localhost:8545')
-const provider = new ethers.providers.Web3Provider(currentProvider);
+let currentProvider;
+let provider;
+
+if (['buidlerevm', 'localhost'].includes(bre.network.name)) {
+  currentProvider = new bre.web3.providers.HttpProvider('http://localhost:8545');
+  provider = new ethers.providers.Web3Provider(currentProvider);
+} else {
+  provider = new ethers.providers.JsonRpcProvider(`https://kovan.infura.io/v3/${INFURA_ID}`);
+}
 
 const waitForTx = async txHash => {
   if (bre.network.name === 'buidlerevm') {
-    return
+    return;
   }
 
-  await provider.waitForTransaction(txHash)
-}
+  console.log('waiting for tx to be mined...', txHash);
+  await provider.waitForTransaction(txHash);
+};
 
 async function main() {
-  const TokenContract = await ethers.getContractFactory("Token");
-  const token = await TokenContract.deploy(config.token.name, config.token.symbol, config.token.totalSupply);
-  await token.deployed();
-  console.log("Token deployed to:", token.address);
+  const {VALIDATOR_PK} = process.env;
 
-  const ValidatorRegistryContract = await ethers.getContractFactory("ValidatorRegistry");
-  const validatorRegistry = await ValidatorRegistryContract.deploy();
-  await validatorRegistry.deployed();
-  console.log("ValidatorRegistry deployed to:", validatorRegistry.address);
-
-  const StakingBankContract = await ethers.getContractFactory("StakingBank");
-  const stakingBank = await StakingBankContract.deploy(token.address, validatorRegistry.address);
-  await stakingBank.deployed();
-  console.log("StakingBank deployed to:", stakingBank.address);
-
-  const ChainContract = await ethers.getContractFactory("Chain");
-  const chain = await ChainContract.deploy(validatorRegistry.address, stakingBank.address, config.chain.interval);
-  await chain.deployed();
-  console.log("Chain deployed to:", chain.address);
-
-  const accounts = await ethers.getSigners();
-
-  for (let index = 0; index < config.validators.length; index++) {
-    const validator = config.validators[index];
-    const id = await accounts[index].getAddress()
-    let tx = await validatorRegistry.create(id, validator.location);
-
-    await waitForTx(tx.hash)
-
-    const validatorData = await validatorRegistry.validators(id)
-    console.log("Added validator number " + index + " with address " + id + " at location " + validatorData.location);
-
-    tx = await token.approve(stakingBank.address, config.token.totalSupply);
-    await waitForTx(tx.hash)
-
-    tx = await stakingBank.receiveApproval(id, config.token.totalSupply, 0);
-    await waitForTx(tx.hash)
-
-    console.log("validator balance:", (await token.balanceOf(id)).toString());
-    console.log("staked balance:", (await stakingBank.balanceOf(id)).toString());
+  if (!VALIDATOR_PK) {
+    throw new Error('please setup VALIDATOR_PK in .env');
   }
 
+  const [owner] = await ethers.getSigners();
+  console.log('DEPLOYING FROM ADDRESS:', await owner.getAddress());
+
+  const validatorWallet = new ethers.Wallet(VALIDATOR_PK, provider);
+  const id = await validatorWallet.getAddress();
+
+  const TokenContract = await ethers.getContractFactory('Token');
+  const token = await TokenContract.deploy(config.token.name, config.token.symbol, config.token.totalSupply);
+  await token.deployed();
+  console.log('Token deployed to:', token.address);
+
+  const ValidatorRegistryContract = await ethers.getContractFactory('ValidatorRegistry');
+  const validatorRegistry = await ValidatorRegistryContract.deploy();
+  await validatorRegistry.deployed();
+  console.log('ValidatorRegistry deployed to:', validatorRegistry.address);
+
+  const StakingBankContract = await ethers.getContractFactory('StakingBank');
+  const stakingBank = await StakingBankContract.deploy(token.address, validatorRegistry.address);
+  await stakingBank.deployed();
+  console.log('StakingBank deployed to:', stakingBank.address);
+
+  const ChainContract = await ethers.getContractFactory('Chain');
+  const chain = await ChainContract.deploy(validatorRegistry.address, stakingBank.address, config.chain.interval);
+  await chain.deployed();
+  console.log('Chain deployed to:', chain.address);
+
+  let tx = await token.transfer(id, config.token.totalSupply);
+  await waitForTx(tx.hash);
+  console.log('token transfered to validator:', config.token.totalSupply);
+
+  if (['buidlerevm', 'localhost'].includes(bre.network.name)) {
+    tx = await owner.sendTransaction({to: id, value: ethers.utils.parseEther('50.0')});
+    await waitForTx(tx.hash);
+  }
+
+  // todo - make it work for multiple validators in a future
+  const validator = config.validators[0];
+
+  tx = await validatorRegistry.create(id, validator.location);
+  await waitForTx(tx.hash);
+
+  const validatorData = await validatorRegistry.validators(id);
+  console.log('Added validator with address ' + id + ' at location ' + validatorData.location);
+
+  tx = await token.connect(validatorWallet).approve(stakingBank.address, config.token.totalSupply);
+  await waitForTx(tx.hash);
+
+  tx = await stakingBank.receiveApproval(id, config.token.totalSupply, 0);
+  await waitForTx(tx.hash);
+
+  console.log('validator balance:', (await token.balanceOf(id)).toString());
+  console.log('staked balance:', (await stakingBank.balanceOf(id)).toString());
+
   const leader = await chain.getLeaderAddress();
-  console.log("Current leader: " + leader);
+  console.log('Current leader: ' + leader);
 }
 
 main()