+++ CHANGELOG.md
@@ -9,7 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - initial version
 - contracts interfaces
 - first class chain data
-
+- proof validation on-chain
+ 
 ### Changed
 - `StakingBank` to be `ERC20` compatible
 - make Chain states: `validatorRegistry`, `stakingBank`, `interval` to be public
+- replace sparse merkle tree with sorted merkle tree